Proyecto 01: Se agregó manejo de excepciones, y completó el CMakeList…

+++ tetris_player/src/include/TetrisSolver/Solver.hpp
@@ -51,7 +51,7 @@ std::vector<PlayState>& Solver::getBestMoves()
 
 bool Solver::findBestMoves(const PlayState& current, int currentDepth)
 {
-	int finalDepth{ gameState.depth };
+	int finalDepth{ static_cast<int>(gameState.depth) };
 	bool isHighScore{false};
 
 	if (currentDepth < finalDepth)
@@ -76,9 +76,9 @@ bool Solver::findChildren(const PlayState& parent, int parentDepth)
 	int currentDepth{ parentDepth + 1 };
 
 	Tetrimino::Figure nextTetrimino{ gameState.nextTetriminos[currentDepth] };
-	int rotations{ Tetrimino::getTetriminoRotations(nextTetrimino) };
+	size_t rotations{ Tetrimino::getTetriminoRotations(nextTetrimino) };
 
-	for (int r{ 0 }; r < rotations; ++r)
+	for (size_t r{ 0 }; r < rotations; ++r)
 	{
 		for (size_t c{ 0 }; c < parent.getPlayArea().cols; ++c)
 		{
